Refactor into service objects
